--- chat-react2/src/components/Home/Sidebar/NewChat/CreateGroupChat/Participant/Participant.scss
@@ -1,10 +1,0 @@
-.group-list {
-  & .group-list__avatar {
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    object-fit: cover;
-  }
-  & .group-username {
-  }
-}
