+++ chat-react2/src/components/Home/Sidebar/NewChat/NewChatDialog.tsx
@@ -8,7 +8,7 @@ import { Button } from "@/components/shadcn-ui/button";
 import { Users } from "node_modules/connectycube/dist/types/types";
 import SearchedUser from "./SearchedUser";
 import CreateGroupChat from "./CreateGroupChat/CreateGroupChat";
-import Participant from "./CreateGroupChat/Participant/Participant";
+import Participant from "./CreateGroupChat/Participant";
 import { useNavigate } from "react-router";
 
 export type ChatType = "private" | "group";
